+++ android/app/src/main/java/com/xu42/tv/live/impl/IBaseBindingPresenter.java
@@ -1,5 +1,0 @@
-package com.xu42.tv.live.impl;
-
-public interface IBaseBindingPresenter {
-
-}